+++ src/layouts/default.vue
@@ -1,10 +1,12 @@
 <template>
   <div class="klab-main-app">
     <template  v-if="!isApp || layout !== null">
+      <klab-notifications></klab-notifications>
       <klab-layout :layout="layout"></klab-layout>
       <app-dialogs></app-dialogs>
       <connection-status class="print-hide"></connection-status>
       <klab-settings class="print-hide"></klab-settings>
+      <klab-terminal v-for="terminal in terminals" :terminal="terminal" :key="terminal.id"></klab-terminal>
       <klab-presentation></klab-presentation>
       <knowledge-view-viewer></knowledge-view-viewer>
     </template>
@@ -33,9 +35,11 @@ import { WEB_CONSTANTS } from 'shared/Constants';
 import { MESSAGES_BUILDERS } from 'shared/MessageBuilders';
 import ConnectionStatus from 'components/ConnectionStatusModal';
 import KlabSettings from 'components/KlabSettings';
+import KlabTerminal from 'components/KlabTerminal';
 import AppDialogs from 'components/AppDialogsViewer';
 import KlabLayout from 'components/KlabLayout.vue';
 import KlabPresentation from 'components/KlabPresentation';
+import KlabNotifications from 'components/KlabNotifications';
 import KnowledgeViewViewer from 'components/KlabKnowledgeViewViewer';
 import 'simplebar/dist/simplebar.css';
 
@@ -45,8 +49,10 @@ export default {
     KlabLayout,
     ConnectionStatus,
     KlabSettings,
+    KlabTerminal,
     AppDialogs,
     KlabPresentation,
+    KlabNotifications,
     KnowledgeViewViewer,
   },
   data() {
@@ -55,6 +61,9 @@ export default {
     };
   },
   computed: {
+    ...mapGetters('data', [
+      'terminals',
+    ]),
     ...mapGetters('view', [
       'layout',
       'isApp',