+++ src/store/data/mutations.js
@@ -2,6 +2,7 @@ import { /* getNodeFromObservation, */findNodeById, WKTInstance } from 'shared/H
 import { eventBus } from 'plugins/initApp';
 import { SCALE_TYPE, SCALE_VALUES, CUSTOM_EVENTS } from 'shared/Constants';
 import { MAP_CONSTANTS } from 'shared/MapConstants';
+import { WEB_CONSTANTS } from '../../shared/Constants';
 // import { DATAFLOW_STATUS } from 'shared/Constants';
 
 export default {
@@ -527,5 +528,28 @@ export default {
 
   SET_CROSSING_IDL: (state, value) => {
     state.crossingIDL = value;
+  },
+
+  ADD_TERMINAL: (state, terminal) => {
+    state.terminals.push(terminal);
+  },
+
+  REMOVE_TERMINAL: (state, terminalId) => {
+    const idx = state.terminals.findIndex(t => t.id === terminalId);
+    if (idx !== -1) {
+      state.terminals.splice(idx, 1);
+    } else {
+      console.warn(`Trying to remove unknown terminal ${terminalId}`);
+    }
+  },
+
+  ADD_TERMINAL_COMMAND: (state, command) => {
+    state.terminalCommands.push(command);
+    localStorage.setItem(WEB_CONSTANTS.LOCAL_STORAGE_TERMINAL_COMMANDS, JSON.stringify(state.terminalCommands));
+  },
+
+  CLEAR_TERMINAL_COMMANDS: (state) => {
+    state.terminalCommands.splice(0, state.terminalCommands.length);
+    localStorage.setItem(WEB_CONSTANTS.LOCAL_STORAGE_TERMINAL_COMMANDS, JSON.stringify(state.terminalCommands));
   },
 };