+++ CHANGELOG.md
@@ -7,10 +7,15 @@ is not available.
 
 **The version of k.Explorer software is self-referred and is not linked to the k.LAB version, so is possible that some features are not available in the stable version of the engine.**
 ## Next
+### Features
+- Notifications from static file
+- Terminal for developers
+
+## 0.12.3
 ### Changes
 - Click on search bar give the focus
 - First implementation of search in app
-- Added upload icon
+- Added upload icon (forthcoming)
 
 ## 0.12.2
 ### Features